add build phases to build-all.sh

+++ script/build-all.sh
@@ -4,6 +4,7 @@ set -x
 
 # Parameters
 
+# PHASE
 # VERSION
 # IMAGE_OUTPUT
 # weatherclock_client_security_apikey
@@ -38,18 +39,35 @@ if [ -z "$VERSION" ] || [ -z "$IMAGE_OUTPUT" ] || [ -z "$weatherclock_client_sec
   exit 1
 fi
 
+if [ -z "$PHASE" ] ; then
+  PHASE="all"
+fi
+
 # Vars for building the client bundle
 export BUNDLE_OUTPUT=$(mktemp -d)
 
 ./script/build-client-bundle.sh
+
+if [ "$PHASE" = "bundle" ] ; then
+  exit 0
+fi
 
 # Get the output filename, should be a .tar.gz file in BUNDLE_OUTPUT_DIR
 export UPLOAD_FILENAME=$(ls $BUNDLE_OUTPUT/*.tar.gz | head -n 1)
 
 ./script/upload-version.sh
 
+if [ "$PHASE" = "upload" ] ; then
+  exit 0
+fi
+
 mkdir -p /opt/timechief-launcher
 rm -rf /opt/timechief-launcher/*
 
 ./script/initialise-launcher-fs.sh
+
+if [ "$PHASE" = "init-fs" ] ; then
+  exit 0
+fi
+
 ./script/build-image.sh
